インプット機能を抜き出してInputクラスを作成、合わせてNoteListのメソッドを変更
