+++ 04.class/lib/note_list.js
@@ -1,36 +1,19 @@
-import readline from "readline";
+// import readline from "readline";
 import Enquirer from "enquirer";
-
+import { Input } from "./input.js";
 export class NoteList {
   constructor(dbConnector) {
     this.dbConnector = dbConnector;
   }
 
   async createNote() {
-    const rl = readline.createInterface({
-      input: process.stdin,
-      output: process.stdout,
-    });
-    const lines = [];
-
+    const input = new Input();
     try {
-      await this.#promiseBasedReadlineOn(rl, "line", (input, resolve) => {
-        if (input === "EOF") {
-          if (lines.length === 0) {
-            rl.close();
-            resolve();
-            return;
-          }
-          if (!lines[0]) {
-            lines[0] = "NoTitle";
-          }
-          this.dbConnector.addNote(lines);
-          rl.close();
-          resolve();
-        } else {
-          lines.push(input);
-        }
-      });
+      const textLines = await input.askNote();
+      if (textLines.length === 0) {
+        return;
+      }
+      this.dbConnector.addNote(textLines);
     } catch (err) {
       if (err instanceof Error && err.code === "SQLITE_ERROR") {
         console.error(err);
@@ -69,13 +52,22 @@ export class NoteList {
     });
   }
 
-  #promiseBasedReadlineOn = (rl, event, callback) => {
-    return new Promise((resolve) => {
-      rl.on(event, (input) => {
-        callback(input, resolve);
-      });
-    });
-  };
+  async #selectNote(messageText, callback) {
+    try {
+      const notes = await this.dbConnector.fetchAllNotes();
+      if (!notes) {
+        return;
+      }
+      const note = await this.#userSelect(notes, messageText);
+      callback(note);
+    } catch (err) {
+      if (err instanceof Error && err.code === "SQLITE_ERROR") {
+        console.error(err);
+      } else {
+        throw err;
+      }
+    }
+  }
 
   async #userSelect(notes, messageText) {
     const question = {
@@ -92,22 +84,5 @@ export class NoteList {
     };
     const answer = await Enquirer.prompt(question);
     return answer.toSeeNote;
-  }
-
-  async #selectNote(messageText, callback) {
-    try {
-      const notes = await this.dbConnector.fetchAllNotes();
-      if (!notes) {
-        return;
-      }
-      const note = await this.#userSelect(notes, messageText);
-      callback(note);
-    } catch (err) {
-      if (err instanceof Error && err.code === "SQLITE_ERROR") {
-        console.error(err);
-      } else {
-        throw err;
-      }
-    }
   }
 }
